requirements.txt
django==3.1.4
djangorestframework==3.12.2
psycopg2==2.8.6
django-fsm==*
